refactor(orchestrator): split actor.go into per-concern files (#410)
actor.go had grown to 2466 lines — the single-authority actor goroutine,
every stateOp implementation, and all reconcile/retry/cleanup machinery in
one file, which the issue flags as costly to review and conflict-prone.

Split it into seven per-concern files in the same package, moving whole
declarations verbatim (no logic change):

  actor.go (376)            core actor goroutine, lifecycle, submit/stateOp, setters
  actor_config.go (309)     config hot-reload + snapshot/query entry points
  actor_dispatch.go (234)   RequestDispatch path, dispatchOp, spawn
  actor_retry.go (597)      RetryScheduler, schedule* entry points, retry stateOps
  actor_finalize.go (254)   finalizeRunOp (terminal RunResult application)
  actor_reconcile.go (510)  Reconcile* entry points + reconcile stateOps
  actor_cleanup.go (246)    reconciled-workspace cleanup machinery

Pure structural move. Verified behavior-preserving:
- top-level declaration set identical before/after (118 == 118);
- go test -race ./internal/orchestrator/ passes;
- funlen/gocognit finding set byte-identical to main (7 == 7, no new
  violations) — the pre-existing oversized apply methods
  (finalizeRunOp.apply, retryFireOp.apply, ...) are unchanged and left for
  a separate, characterization-test-backed decomposition PR.

Addresses #410 (actor.go file-size hotspot).
